+++ my_game_list/games/admin/game_review.py
@@ -11,8 +11,5 @@ class GameReviewAdmin(admin.ModelAdmin[GameReview]):
     readonly_fields = ("id",)
     search_fields = (*readonly_fields, "game__title", "user__username")
     raw_id_fields = ("game", "user")
-    list_filter = (
-        "score",
-        "created_at",
-    )
+    list_filter = ("created_at",)
     list_display = (*readonly_fields, *list_filter, *raw_id_fields)