+++ my_game_list/users/models/__init__.py
@@ -1,6 +1,7 @@
 """This package contains all the models for the users application."""
-from my_game_list.users.models.user import User
+from my_game_list.users.models.user import Gender, User
 
 __all__ = [
+    "Gender",
     "User",
 ]