+++ my_game_list/users/serializers/__init__.py
@@ -1,7 +1,8 @@
 """This package contains all the serializers for the users application."""
-from my_game_list.users.serializers.user import UserCreateSerializer, UserSerializer
+from my_game_list.users.serializers.user import UserCreateSerializer, UserDetailSerializer, UserSerializer
 
 __all__ = [
     "UserCreateSerializer",
+    "UserDetailSerializer",
     "UserSerializer",
 ]