+++ my_game_list/users/serializers/user.py
@@ -4,7 +4,11 @@ from typing import TYPE_CHECKING, Any, ClassVar, Self
 
 from django.contrib.auth import get_user_model
 from django.contrib.auth.password_validation import validate_password as django_validate_password
+from django.db.models import Avg
 from rest_framework import serializers
+from rest_framework.utils.serializer_helpers import ReturnDict
+
+from my_game_list.games.models import GameListStatus
 
 if TYPE_CHECKING:
     from my_game_list.users.models import User as UserType
@@ -19,7 +23,7 @@ class UserCreateSerializer(serializers.ModelSerializer["UserType"]):
         """Meta data for the class."""
 
         model = User
-        fields = ("username", "password", "email", "avatar")
+        fields = ("username", "password", "email", "avatar", "gender")
         extra_kwargs: ClassVar[dict[str, dict[str, bool]]] = {"password": {"write_only": True}}
 
     def validate_password(self: Self, value: str) -> str:
@@ -39,8 +43,20 @@ class UserCreateSerializer(serializers.ModelSerializer["UserType"]):
         return user
 
 
+class UserSimpleSerializer(serializers.ModelSerializer["UserType"]):
+    """Simple user serializer with only avatar and url to the details."""
+
+    class Meta:
+        """Meta data for the class."""
+
+        model = User
+        fields = ("id", "avatar")
+
+
 class UserSerializer(serializers.ModelSerializer["UserType"]):
     """Serializer for listing the user model."""
+
+    gender = serializers.CharField(source="get_gender_display", read_only=True)
 
     class Meta:
         """Meta data for the class."""
@@ -50,9 +66,59 @@ class UserSerializer(serializers.ModelSerializer["UserType"]):
             "id",
             "username",
             "email",
+            "gender",
             "last_login",
             "date_joined",
             "avatar",
             "is_active",
         )
         read_only_fields = ("id", "avatar", "last_login", "date_joined")
+
+
+class UserDetailSerializer(serializers.ModelSerializer["UserType"]):
+    """Detailed serializer for User model."""
+
+    gender = serializers.CharField(source="get_gender_display", read_only=True)
+    game_list_statistics = serializers.SerializerMethodField()
+    friends = serializers.SerializerMethodField()
+    latest_game_list_updates = serializers.SerializerMethodField()
+
+    class Meta:
+        """Meta data for the class."""
+
+        model = User
+        fields = (
+            "id",
+            "username",
+            "gender",
+            "last_login",
+            "date_joined",
+            "avatar",
+            "game_list_statistics",
+            "friends",
+            "latest_game_list_updates",
+        )
+
+    def get_game_list_statistics(self: Self, instance: "UserType") -> dict[str, int]:
+        """Get the game list statistics for the user."""
+        return {
+            "completed": instance.game_lists.filter(status=GameListStatus.COMPLETED).count(),
+            "dropped": instance.game_lists.filter(status=GameListStatus.DROPPED).count(),
+            "plan_to_play": instance.game_lists.filter(status=GameListStatus.PLAN_TO_PLAY).count(),
+            "on_hold": instance.game_lists.filter(status=GameListStatus.ON_HOLD).count(),
+            "playing": instance.game_lists.filter(status=GameListStatus.PLAYING).count(),
+            "total": instance.game_lists.count(),
+            "mean_score": instance.game_lists.aggregate(mean_score=Avg("score"))["mean_score"],
+        }
+
+    def get_friends(self: Self, instance: "UserType") -> ReturnDict[Any, Any]:
+        """Get the list of friends for the user."""
+        friendships = instance.friends.all()[:5]
+        friends = [friendship.friend for friendship in friendships]
+        return UserSimpleSerializer(friends, many=True).data
+
+    def get_latest_game_list_updates(self: Self, instance: "UserType") -> ReturnDict[Any, Any]:
+        """Get the latest game list updates for the user."""
+        from my_game_list.games.serializers.game_list import GameListSerializer
+
+        return GameListSerializer(instance.game_lists.all().order_by("last_modified_at")[:5], many=True).data